feat: Restart Thread

+++ Modules/ChatGPTModule.cs
@@ -6,9 +6,6 @@ using OpenAI.ObjectModels.RequestModels;
 using OpenAI.ObjectModels;
 using OpenAI.Managers;
 using DiscordBot.Utils;
-using AngleSharp.Browser;
-using YoutubeExplode.Videos.ClosedCaptions;
-using System.Text.Unicode;
 using System.Text;
 using DiscordBot.Domain.Entities;
 
@@ -117,7 +114,7 @@ namespace DiscordBot.Modules
 
             await DeferAsync();
 
-           var channel = await ((ITextChannel)Context.Channel).CreateThreadAsync(nome, ThreadType.PublicThread);
+            var channel = await ((ITextChannel)Context.Channel).CreateThreadAsync(nome, ThreadType.PublicThread);
 
             await db.threads.UpsertAsync(new ThreadChannel
             {
@@ -126,6 +123,33 @@ namespace DiscordBot.Modules
                 MemberId = Context.Client.CurrentUser.Id,
                 LastMessageTime = channel.CreatedAt.UtcDateTime,
             });
+            await ((ISlashCommandInteraction)Context.Interaction).DeleteOriginalResponseAsync();
+        }
+
+        [SlashCommand("restart-thread", "Apagar as msgs de uma tread privada com o Ademir.")]
+        public async Task RestartThread()
+        {
+            IThreadChannel? ch = Context.Channel as IThreadChannel;
+            var me = await Context.Guild.GetUserAsync(Context.User.Id);
+            if(ch == null)
+            {
+                await RespondAsync("Você não está em uma thread com o Ademir.");
+                return;
+            }
+
+            var msgs = (await ch.GetMessagesAsync(ch.MessageCount).FlattenAsync()).OrderBy(m => m.Timestamp);
+            if(msgs.FirstOrDefault()?.Author.Id != Context.User.Id && !me.GuildPermissions.Administrator)
+            {
+                await RespondAsync("Você precisa ser o solicitante da thread.");
+                return;
+            }
+
+            await DeferAsync();
+            foreach (var msg in msgs)
+            {
+                await ch.DeleteMessageAsync(msg.Id);
+            }
+
             await ((ISlashCommandInteraction)Context.Interaction).DeleteOriginalResponseAsync();
         }
     }
